Connected

+++ src/app/api/posts/[id]/route.ts
@@ -1,5 +1,7 @@
 import { NextResponse } from "next/server";
 import prisma from "@/app/lib/prismadb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../auth/[...nextauth]/route";
 export async function GET(
   req: Request,
   { params }: { params: { id: string } }
@@ -18,7 +20,10 @@ export async function GET(
 }
 
 export async function PUT(req:Request,{params}:{params:{id:string}}) {
-
+  const session = await getServerSession(authOptions);
+  if(!session){
+    return NextResponse.json({error:"Not authenticated"},{status:401})
+  }
     const {title, content,links,selectedCategory,imageUrl,publicId} = await req.json()
     const id = params.id 
     try {
@@ -44,6 +49,10 @@ export async function PUT(req:Request,{params}:{params:{id:string}}) {
 }
 
 export async function DELETE(req:Request, {params}:{params:{id:string}}){
+  const session = await getServerSession(authOptions);
+  if(!session){
+    return NextResponse.json({error:"Not authenticated"},{status:401})
+  }
   const id = params.id;
   try {
     const post = await prisma.post.delete({where:{id}})
